refactor editor node details into dedicated components
